fix(iconpack): close atomic publication platform gaps
